Janela modal, redirecionamento e não repetir email

+++ codigo-fonte/assets/js/cadastro.js
@@ -49,6 +49,25 @@ function passwordV() {
     }
 }
 
+function successMsg() {
+
+    const cadastro = document.querySelector('#cadastro')
+    
+    const modalWindow = document.createElement('div');
+    modalWindow.classList.add('window');
+    
+    const modalInterior = document.createElement('div');
+    modalInterior.classList.add('interior');
+    
+    const modalMessage = document.createElement('p');
+    modalMessage.textContent = "Cadastro feito com sucesso!";
+                                
+    modalWindow.append(modalInterior, modalMessage);
+    cadastro.append(modalWindow);
+
+    setTimeout(function() {modalWindow.remove(); window.location.href = "curso.html";}, 1000);
+}
+
 form.addEventListener("submit", function(cadastro) {
     cadastro.preventDefault();
     
@@ -65,14 +84,33 @@ form.addEventListener("submit", function(cadastro) {
         users = JSON.parse(localStorage.getItem("users"));
         console.log({users});
     }
-    
+
     const user = {
         id: newId,
         name: name.value, 
         email: email.value, 
         password: password.value
     }
+
+    const userEmail = email.value;
+    let exist = false;
+
+    for(let i = 0; i < users.length; i++) {
+        if(users[i].email === userEmail) {
+            exist = true
+            break;
+        }
+    } 
+
+    if(exist) {
+        errorMessage.style.display = 'block';
+        errorMessage.textContent = 'Este e-mail já está cadastrado.';
+        return;
+    }
+    
     users.push(user);
     
     localStorage.setItem("users", JSON.stringify(users));
-})
+
+    successMsg();  
+});
